Revert "Extract callback ToCallbackList mapping method to InboundMapper class"
This reverts commit fbbe8db57d741b4169b8454f85e7f4f38446d1d9.
